Improve XHR status messages and page validation
Use a loading counter instead of a boolean value to accomodate for overlapping XHRs. This also sets the loading state correctly when the viewer is opened. On error, distinguish between the network being down and the server sending an invalid response.

Set the page to 1 if an invalid value is provided (which requires the manifest being loaded first), and add a unit test.