Improve scan behavior
If the image was panned or zoomed on page turn, pan to ensure the upper-left corner is visible since this is the most likely part in which the user is interested in. Otherwise, reset the view.

In double-page view, move the first page to the right and the last page - if even - to the left.

Show the loading indicator until all tiles have been loaded and ensure the pan and zoom parameters are removed when the scan is reset.