Display metadata of active structure (GDZ-641)
This selection bases on the idea that the current element's page range
is the smallest range of all the elements that contain the current canvas.
Other elements containing the current canvas, but not being the current
element might be higher ranging elements that should (hopefully) span
more pages.

Add test to check selection of active structure