Fullscreen functionality (#33)
* Add fullscreen button

Add fullscreen button HTML to .tify-scan section. The fullscreen button will be conditionally rendered depending on whether the browser supports the fullscreen API. If so, the fullscreen button icon  will change depending on whether fullscreen has been activated or not.

* Add fullscreen properties to the data object

The fullscreenActive property detects whether fullscreen has been activated. It is false by default. The screen property grabs the DOM element that will be made fullscreen when fullscreen mode is activated.

* Add detectFullscreen() function to the methods object

The detectFullscreen() function returns the appropriate fullscreenElement property depending on the browser. Otherwise, it returns false (i.e. the browser does not support the fullscreen API and therefore the fullscreen button will not be rendered).

* Add toggleFullscreen() function to the methods object

The toggleFullscreen() function activates and deactivates fullscreen mode. It also toggles the fullscreenActive state in the data object, which will cause the fullscreen button icon to change.

* Assign 'u'/'U' to toggle fullscreen

* Add French translation for fullscreen/exit fullscreen

* Add German translation for fullscreen/exit fullscreen

* Add E2E tests

Add tests to check that fullscreen mode can be toggled using the mouse and keyboard

* Updated English and German key bindings

* Modified screen property to use options.container

* Fixed spacing

* Implement fullscreen functionality in Header.vue

Move fullscreen buttons and methods from Scan.vue to Header.vue

* Add comments to detectFullscreen()