Use Github Actions for Testing on Linux, Mac and Windows
Github Actions provide Runners for Linux, Mac and Windows to test Applications.
We can omit Travis and Appveyor in favour of using Github Actions.
As it was before, on Windows only the unit tests are executed - Fixing this is
not part of this Commit.
Also the Fullscreen e2e Tests are failing, that may also be fixed later, as the
functionality is working on a 'real' computer.
Updating the Testing-Framework also was necessary.