Fix button not closing popup, reduce redundancy
Since v-click-outside trumps touch listeners, we need to move the directive to a common parent element and change the DOM tree accordingly. Otherwise, the button does not close the popup on touch. Extend associated e2e tests.

The pagination buttons' DOM is now identical for large and small screens, so move them to a separate component.