Remove canvas package and add sass
The package does not seem to be in use and tests are working without the
package.
Installing canvas relies on native bindings for the operating system and
e.g. on ARM-based Apple computers workarounds are needed.
Also sass is required to compile the sass-based stylesheets. The
requirement has been added.